requirements.txt
numpy
scipy
torch
torchaudio
transformers>=4.36.0
librosa
soundfile
pillow
pillow-avif-plugin
